Introdotte response

+++ code/dietiestates-backend/src/main/java/com/dietiestates/api/dto/response/CadastralDataResponse.java
@@ -1,5 +1,6 @@
 package com.dietiestates.api.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.Builder;
@@ -14,12 +15,12 @@ import lombok.ToString;
 public class CadastralDataResponse {
 
 	private Long id;
-	private Integer price;
+	private BigDecimal price;
 	private Integer squareMeters;
 	private String energyClass;
 	private Integer rooms;
 	private Integer floor;
 	private LocalDateTime createdDate;
 	private LocalDateTime lastModifiedDate;
-	private Integer realEstateId;
+	private Long realEstateId;
 }
