+++ code/dietiestates-backend/src/main/java/com/dietiestates/api/dto/response/NotificationResponse.java
@@ -17,6 +17,6 @@ public class NotificationResponse {
 	private LocalDateTime createdDate;
 	private LocalDateTime lastModifiedDate;
 	private String message;
-	private Integer notificationCategoryId;
+	private Long notificationCategoryId;
 	private String userEmail;
 }
